service/debugger: merge attachErrorMessage, verifyBinaryFormat
These functions don't need to be conditionally compiled and they share
substanatial amounts of code in all their OS-specific versions.
